==> dataset_assets/preprocess/3_channel_data_clean.ipynb <==
import pandas as pd
import json
import os

For the channel videos dataset (after combination), we will commence the similar cleaning techniques as used in trending dataset (percentile-based), but this time, only 10% will be cut off (those videos won't likely be trending as of Dec.2024)

channel_file_path = "C:/Users/TKN/Downloads/New-Youtube-Scraper-v3/data/combined_channel_data.json"
with open(channel_file_path, "r", encoding="utf-8") as f:
    trending_videos = json.load(f)

channel_df = pd.DataFrame.from_dict(trending_videos, orient="index")

channel_df.info()

`view_stat` function to view the lowest numbers across the vital categories.

def view_stats(column_name):
    list = sorted(channel_df[column_name].tolist())
    print(list[:10])

view_stats('viewCount')
view_stats('likeCount')
view_stats('commentCount')
view_stats('engagementRate')
view_stats('avgDailyViews')

`view_lowest` function to view the videos with the lowest interactions (view count, like count, comment count, average daily views)

def view_lowest(column_name):
    lowest_view_count_row = channel_df.loc[channel_df[column_name].idxmin()]
    print(f"Title: {lowest_view_count_row['title']}")
    print(f'Published At: {lowest_view_count_row["publishedAt"]}')
    print(f"View Count: {lowest_view_count_row['viewCount']}")
    print(f"Like Count: {lowest_view_count_row['likeCount']}")
    print(f"Comment Count: {lowest_view_count_row['commentCount']}")
    print(f"Avg Daily Views: {lowest_view_count_row['avgDailyViews']}")
    print(f"Engagement Rate: {lowest_view_count_row['engagementRate']}")

view_lowest('viewCount')

percentiles = channel_df[['viewCount', 'likeCount', 'commentCount', 'engagementRate', 'avgDailyViews']].quantile(0.1)
print("10th percentiles:")
print(percentiles)

We take rounded numbers for the filtered dataset (keep the original numbers won't change number of data points). Only the videos which satisfy all the conditions remain in the dataset.

We lower the `engagementRate` for channel video dataset this time since lots of videos (music videos and shorts mostly) have a low `engagementRate`. 0.001 is the industry-standardized amount for engagement rate of large channels (around 1-10 million views)

channel_df = channel_df[
    (channel_df["viewCount"] >= 90000)
    & (channel_df['likeCount'] >= 2300)
    & (channel_df['commentCount'] >= 70)
    & (channel_df['engagementRate'] >= 0.001)
    & (channel_df['avgDailyViews'] >= 170)
]

channel_df.shape[0]

#### Assigning trending labels

Similar to the trending dataset, we will also assign `isTrending` and `trendingPercentile`. But we need to assign `isTrending` first. The result is a dataset with `isTrending` either `1` or `0`. 

The videos with `1` (they are trending) will be separated from videos with `0`. 

Videos with `1` then get concatenated with trending videos to assign `trendingPercentile` to serve for prediction

Videos with `0` will also get a `nonTrendingPercentile` ? With top means (not trending, still significance) and bottom means (not trending, little significance)

From trending dataset, we obtained the benchmark for trending videos is as follow
```
Trending Benchmarks:
   isTrending  avgDailyViews  viewCount  likeCount  commentCount
0           1      102733.75     397425       9619          1086
```
The videos which satisfy all these conditions will be marked as trending (1) else (0)

trending_benchmarks = {
    'avgDailyViews': 102733.75,
    'viewCount': 397425,
    'likeCount': 9619,
    'commentCount': 1086
}

channel_df['isTrending'] = (
    (channel_df['avgDailyViews'] >= trending_benchmarks['avgDailyViews']) &
    (channel_df['viewCount'] >= trending_benchmarks['viewCount']) &
    (channel_df['likeCount'] >= trending_benchmarks['likeCount']) &
    (channel_df['commentCount'] >= trending_benchmarks['commentCount'])
).astype(int)

trending_count_df = channel_df[channel_df['isTrending'] == 1]

trending_count_df.shape[0] # videos in channel video dataset is marked as trending

channel_df_rep = channel_df[['publishedAt','elapsedDays', 'title', 'channelTitle', 
                           'category','topicCategories', 'duration', 'licensedContent',
                           'viewCount', 'likeCount', 'commentCount', 'avgDailyViews',
                           'engagementRate', 'isTrending']]
channel_df_rep.info()

output_dir_2 = 'C:/Users/TKN/Downloads/New-Youtube-Scraper-v3/data/yt_processed_data'
combined_json = channel_df.to_json(orient="index", force_ascii=False, indent=4)

with open(os.path.join(output_dir_2, "processed_channel_videos.json"), "w", encoding="utf-8") as f:
    f.write(combined_json)    
print(f"Combined channel_data JSON saved to {output_dir_2}.")
